Remove RecentsViewModel.kt
